Introduce and use overlay at all levels
Now that we have an "asahi" overlay including definitions for all of our
packages, let's use it for both `nixos-m1` (installer ISO etc) and within
`m1-support`. In the latter case, all `pkgs` args are presumed to now
have the overlay applied, so packages can cross-reference each other.

Finally, because of `override` and lazy eval, in no way will this
impede custom configuration of any package going forward. This may be
obvious to nix veterans, but it wasn't to me until it was. :)