replace system Mesa when using experimental GPU support
This avoids breaking programs like Plasma Wayland due to either the slight
Mesa version mismatch or the adjustments made for Asahi GPU support.

`system.replaceRuntimeDependencies` is used instead of an overlay to avoid
rebuilding the world for this replacement. It's possible this might cause
its own issues but this is all experimental anyway.