peripheral-firmware: extract on boot by default
This improves the situation for flake users, as previously they had to
either add the non-redistributable firmware to their repositories or go
without the firmware.

This change also brings the ISO into line with the regular NixOS modules
as storing the peripheral firmware in the Nix Store is no longer a
requirement.