linux-asahi: update kernel config to 6.6.x format
This update makes no functional changes to the kernel config.  It is merely the result of running the latest nixos-apple-silicon config through 'make olddefconfig'.

This has the result of sorting all options into Kbuild/Kconfig order, and adding in entries for all config options introduced between 6.5 and 6.6 (with their values set to the defaults defined in the arm64 defconfig).

This commit also makes the edge kernel config the default, for two reasons:
1) the Fedora Asahi Remix enables the GPU and DCP drivers by default, and doesn't make an edge/non-edge distinction,
2) the nixos-apple-silicon project already did so (inadvertently?) when they added DRM_ASAHI and DRM_APPLE to the baseline configuration in the commit "linux-asahi: sync Apple Silicon specific options with Fedora Asahi Remix".

Tested by booting successfully on an M1 MacBook Air.