kernel/linux-asahi: build most Apple Silicon-specific drivers into the kernel directly
We can eliminate the possibility of a bunch of timing race condition bugs by making sure as many Apple Silicon drivers are built directly into the kernel image as possible, instead of as modules.  Certain drivers are left as modules because they load firmware, or because I've read reports on IRC that they should be built as modules for the time being.

We then modify the list of available kernel modules to match the changes we've made to the kernel config.