mesa-asahi-edge: un-vendor and override nixpkgs version
Nixpkgs's derivation is now much more viable to use with the libgbm
changes, and it's what they suggest.

It's close in difficulty to only build our drivers vs. add our drivers
to the build, and the former saves a decent amount of compile time
(though not dependencies as we don't modify the buildInputs), so that's
what we do.

* remove Meson flags no longer present in 25.0.0

* create output we don't build but is still in the upstream list