mesa: only install driver now that libgbm is fixed
libgbm is in effect dynamically loaded by applications as of
https://github.com/NixOS/nixpkgs/issues/269419 so we no longer create a
verison mismatch by installing our different Mesa as the driver.
Therefore patching applications is no longer necessary.

Keep the option around a bit in case it turns out this is untrue and we
need to bring back the different patching modes.